New: - Array.randomized() - HTTPCookieStorage.removeAllCookies(for:) - (UI|NS)View.animation - get a set of animations on the views. - XUBorderlessTextField - a text field similar as on Android, with the placeholder displayed above text, with character counter, etc. - XUScrollViewKeyboardInsetManager - a helper class that insets UIScrollView insets as the keyboard is displayed.
Compatibility fixes with Swift 3.2, Xcode 9 and macOS 10.13 SDK.

+++ XUCore/misc/XUTimeUtilities.swift
@@ -25,6 +25,7 @@ public struct XUTime {
 	
 	/// Returns seconds as human-readable string. E.g. 1 hour 10 minutes 1 second.
 	public static func localizedTimeString(_ seconds: TimeInterval) -> String {
+		
 		if seconds < 0 || !seconds.isFinite || TimeInterval(Int64.max) < seconds {
 			return XULocalizedString("1 second", inBundle: XUCoreFramework.bundle)
 		}
@@ -33,7 +34,7 @@ public struct XUTime {
 		var minuteString = ""
 		var secondsString = ""
 		
-		var eta = Int(seconds)
+		var eta = Int64(seconds)
 		if eta > 3599 {
 			// i.e. at least one hour
 			let hours = eta / 3600
